// Front-End/src/app/models/vehicle.exit.ts
export class VehicleExit {
    companyId: number | null = null;
    resaleId: number | null = null;
    vehicleId: number | null = null;
    exitUserId: number | null = null;
    exitUserName: string = '';
    exitDate: string | Date = '';
    exitInformation: string = '';
    exitPhoto1Url: string = '';
    exitPhoto2Url: string = '';
    exitPhoto3Url: string = '';
    exitPhoto4Url: string = '';
}